refactor: Car의 이름 vo로 대체, getName 삭제
