+++ src/main/java/racingCar/domain/Car.java
@@ -3,16 +3,13 @@ package racingCar.domain;
 import racingCar.dto.CarDto;
 
 public class Car implements Comparable<Car> {
-
     private static final int MOVE_MIN_NUM  = 4;
-    private static final int NAME_MAX_LENGTH = 5;
 
-    private String name;
-    private Position position;
+    private final CarName name;
+    private final Position position;
 
     public Car(String carName) {
-        validateLength(carName);
-        this.name = carName;
+        this.name = new CarName(carName);
         position = new Position();
     }
 
@@ -20,19 +17,9 @@ public class Car implements Comparable<Car> {
         return position.getPosition();
     }
 
-    public String getName() {
-        return name;
-    }
-
     public void move(int pickedNumber) {
         if (pickedNumber >= MOVE_MIN_NUM) {
             position.increase();
-        }
-    }
-
-    private void validateLength(String carName) {
-        if (carName.length() > NAME_MAX_LENGTH) {
-            throw new IllegalArgumentException("자동차 이름은 5자 이하만 가능합니다.");
         }
     }
 
@@ -42,7 +29,7 @@ public class Car implements Comparable<Car> {
     }
 
     public CarDto toCarDto() {
-        return new CarDto(this.name, this.position);
+        return new CarDto(this.name.getCarName(), this.position.getPosition());
     }
 
 }
