+++ src/main/java/racingCar/domain/CarName.java
@@ -1,0 +1,23 @@
+package racingCar.domain;
+
+public class CarName {
+    private static final int NAME_MAX_LENGTH = 5;
+
+    private final String carName;
+
+    public CarName(String carName) {
+        this.carName = carName;
+        validateLength(carName);
+    }
+
+    public String getCarName() {
+        return carName;
+    }
+
+    private void validateLength(String carName) {
+        if (carName.length() > NAME_MAX_LENGTH) {
+            throw new IllegalArgumentException("자동차 이름은 5자 이하만 가능합니다.");
+        }
+    }
+
+}
