+++ src/main/java/racingCar/dto/CarDto.java
@@ -1,21 +1,23 @@
 package racingCar.dto;
 
-import racingCar.domain.Position;
-
 public class CarDto {
     private final String name;
-    private final Position position;
+    private final int position;
 
-    public CarDto(String name, Position position) {
+    public CarDto(String name, int position) {
         this.name = name;
         this.position = position;
+    }
+
+    public String getName() {
+        return name;
     }
 
     @Override
     public String toString() {
         StringBuilder stringBuilder = new StringBuilder();
         stringBuilder.append(name + " : ");
-        for (int i = 0; i < position.getPosition(); i++) {
+        for (int i = 0; i < position; i++) {
             stringBuilder.append("-");
         }
         return stringBuilder.toString();
